refactor: change the structure of repository
