--- person_trackers/bytetrack.yaml
@@ -1,7 +1,0 @@
-tracker_type: bytetrack
-track_high_thresh: 0.5
-track_low_thresh: 0.1
-new_track_thresh: 0.6
-track_buffer: 30
-match_thresh: 0.9
-fuse_score: true
